fix: fixed the bug that the style cache was not updated after the permissions were loaded
