+++ packages/sheets-ui/src/controllers/permission/sheet-permission-interceptor-base.controller.ts
@@ -41,7 +41,7 @@ type ICheckPermissionCommandParams = IEditorBridgeServiceVisibleParam | IMoveRow
 
 export const SHEET_PERMISSION_PASTE_PLUGIN = 'SHEET_PERMISSION_PASTE_PLUGIN';
 
-@OnLifecycle(LifecycleStages.Steady, SheetPermissionInterceptorBaseController)
+@OnLifecycle(LifecycleStages.Rendered, SheetPermissionInterceptorBaseController)
 export class SheetPermissionInterceptorBaseController extends Disposable {
     disposableCollection = new DisposableCollection();
