+++ packages/sheets-ui/src/controllers/permission/sheet-permission-render.controller.ts
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-import { Disposable, Inject, IPermissionService } from '@univerjs/core';
+import { Disposable, Inject, IPermissionService, IUniverInstanceService, UniverInstanceType } from '@univerjs/core';
 import type { MenuConfig } from '@univerjs/ui';
 import { ComponentManager } from '@univerjs/ui';
 import { CheckMarkSingle, DeleteSingle, LockSingle, ProtectSingle, WriteSingle } from '@univerjs/icons';
-import { RangeProtectionRuleModel } from '@univerjs/sheets';
+import { RangeProtectionRuleModel, WorksheetProtectionRuleModel } from '@univerjs/sheets';
 import type { IRenderContext, IRenderModule, Spreadsheet } from '@univerjs/engine-render';
+import { IRenderManagerService } from '@univerjs/engine-render';
 import { merge, throttleTime } from 'rxjs';
 import { permissionCheckIconKey, permissionDeleteIconKey, permissionEditIconKey, permissionLockIconKey, permissionMenuIconKey, UNIVER_SHEET_PERMISSION_DIALOG, UNIVER_SHEET_PERMISSION_PANEL, UNIVER_SHEET_PERMISSION_PANEL_FOOTER, UNIVER_SHEET_PERMISSION_USER_DIALOG } from '../../consts/permission';
 import { SheetPermissionDialog, SheetPermissionPanel, SheetPermissionPanelFooter, SheetPermissionUserDialog } from '../../views/permission';
@@ -27,6 +28,7 @@ import { UNIVER_SHEET_PERMISSION_ALERT_DIALOG } from '../../views/permission/err
 import { AlertDialog } from '../../views/permission/error-msg-dialog';
 import { SheetSkeletonManagerService } from '../../services/sheet-skeleton-manager.service';
 import { RANGE_PROTECTION_CAN_NOT_VIEW_RENDER_EXTENSION_KEY, RANGE_PROTECTION_CAN_VIEW_RENDER_EXTENSION_KEY, RangeProtectionCanNotViewRenderExtension, RangeProtectionCanViewRenderExtension } from '../../views/permission/extensions/range-protection.render';
+import { worksheetProtectionKey, WorksheetProtectionRenderExtension } from '../../views/permission/extensions/worksheet-permission.render';
 
 export interface IUniverSheetsPermissionMenuConfig {
     menu: MenuConfig;
@@ -108,7 +110,54 @@ export class SheetPermissionRenderController extends Disposable implements IRend
 
         this.disposeWithMe(merge(
             this._permissionService.permissionPointUpdate$.pipe(throttleTime(300, undefined, { trailing: true })),
+            this._rangeProtectionRuleModel.rangeRuleInitStateChange$,
             this._rangeProtectionRuleModel.ruleChange$
+        ).pipe().subscribe(markDirtySkeleton));
+    }
+}
+
+export class WorksheetProtectionRenderController extends Disposable implements IRenderModule {
+    private _worksheetProtectionRenderExtension = new WorksheetProtectionRenderExtension();
+    constructor(
+        private readonly _context: IRenderContext,
+        @Inject(IRenderManagerService) private _renderManagerService: IRenderManagerService,
+        @Inject(IUniverInstanceService) private _univerInstanceService: IUniverInstanceService,
+        @Inject(SheetSkeletonManagerService) private _sheetSkeletonManagerService: SheetSkeletonManagerService,
+        @Inject(WorksheetProtectionRuleModel) private _worksheetProtectionRuleModel: WorksheetProtectionRuleModel
+
+    ) {
+        super();
+        this._initRender();
+        this._initSkeleton();
+    }
+
+    private _initRender() {
+        const register = (renderId: string) => {
+            const render = renderId && this._renderManagerService.getRenderById(renderId);
+            const spreadsheetRender = render && render.mainComponent as Spreadsheet;
+            if (spreadsheetRender) {
+                if (!spreadsheetRender.getExtensionByKey(worksheetProtectionKey)) {
+                    spreadsheetRender.register(this._worksheetProtectionRenderExtension);
+                }
+            }
+        };
+        this.disposeWithMe(this._renderManagerService.currentRender$.subscribe((renderId) => {
+            renderId && register(renderId);
+        }));
+        const workbook = this._univerInstanceService.getCurrentUnitForType(UniverInstanceType.UNIVER_SHEET)!;
+        if (workbook) {
+            register(workbook.getUnitId());
+        }
+    }
+
+    private _initSkeleton(): void {
+        const markDirtySkeleton = () => {
+            this._sheetSkeletonManagerService.reCalculate();
+            this._context.mainComponent?.makeDirty();
+        };
+
+        this.disposeWithMe(merge(
+            this._worksheetProtectionRuleModel.worksheetRuleInitStateChange$
         ).pipe().subscribe(markDirtySkeleton));
     }
 }
